pyproject.toml
[project]
name = "health-llm-chatbot"
version = "0.1.0"
description = "A friendly health query chatbot powered by LLMs with safety features"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "streamlit>=1.28.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "google-genai>=1.56.0",
]
